+++ image-categorizer.sln
@@ -5,8 +5,6 @@ VisualStudioVersion = 17.3.32825.248
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "image-categorizer", "image-categorizer\image-categorizer.csproj", "{801E96C2-2BB6-4F65-BB70-2B4F4CFD569D}"
 EndProject
-Project("{54435603-DBB4-11D2-8724-00A0C9A8B90C}") = "image-categorizer-setup", "image-categorizer-setup\image-categorizer-setup.vdproj", "{34752A7C-E0B9-4A19-8620-E973216DE08B}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -17,9 +15,6 @@ Global
 		{801E96C2-2BB6-4F65-BB70-2B4F4CFD569D}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{801E96C2-2BB6-4F65-BB70-2B4F4CFD569D}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{801E96C2-2BB6-4F65-BB70-2B4F4CFD569D}.Release|Any CPU.Build.0 = Release|Any CPU
-		{34752A7C-E0B9-4A19-8620-E973216DE08B}.Debug|Any CPU.ActiveCfg = Release
-		{34752A7C-E0B9-4A19-8620-E973216DE08B}.Release|Any CPU.ActiveCfg = Release
-		{34752A7C-E0B9-4A19-8620-E973216DE08B}.Release|Any CPU.Build.0 = Release
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
