+++ image-categorizer/MVVM/Model/InsertQueryModel.cs
@@ -8,16 +8,16 @@ namespace image_categorizer.MVVM.Model
 {
     public class InsertQueryModel
     {
-        public string fileOutputPath;
+        public string fileName;
         public string dateTime;
         public string format;
         public string cameraModel;
         public string location;
         public string currentTime;
 
-        public InsertQueryModel(string fileOutputPath, string dateTime, string format, string cameraModel, string location, string currentTime)
+        public InsertQueryModel(string fileName, string dateTime, string format, string cameraModel, string location, string currentTime)
         {
-            this.fileOutputPath = fileOutputPath;
+            this.fileName = fileName;
             this.dateTime = dateTime;
             this.format = format;
             this.cameraModel = cameraModel = (cameraModel == null || cameraModel == "") ? "No Camera Data" : this.cameraModel = cameraModel;
