+++ image-categorizer/MVVM/View/RunView.xaml
@@ -76,7 +76,7 @@
                             Background="{StaticResource BackgroundSolidColorBrush}"
                             CornerRadius="5">
                             <ProgressBar
-                                Maximum="{Binding RunModel.MaxProgress}"
+                                Maximum="{Binding RunModel.FileCount}"
                                 Minimum="0"
                                 Value="{Binding RunModel.CategorizeProgress, Mode=OneWay}" />
                         </Border>
@@ -92,11 +92,8 @@
                     Content="Run"
                     DockPanel.Dock="Right"
                     FontSize="20"
-                    IsEnabled="{Binding RunModel.IsIdle}"
-                    Foreground="{StaticResource WhiteSolidColorBrush}"
-                    Style="{StaticResource MainAcceptButtonThemes}"
-                    ToolTip="Categorize With Setting Tap's Settings"
-                    ToolTipService.Placement="Left" />
+                    Foreground="{StaticResource WhiteSolidColorBrush}" />
+
             </DockPanel>
         </StackPanel>
     </Border>
