--- image-categorizer/Utility/IGeoCoding.cs
@@ -1,8 +1,0 @@
-﻿namespace image_categorizer
-{
-    public interface IGeoCoding
-    {
-        void GeoCodingInit();
-        string? GetLocation(double? latitude, double? longitude);
-    }
-}
